c端秒杀订单列表、详情

+++ store-saas-marketing-facade/src/main/java/com/tuhu/store/saas/marketing/response/seckill/SeckillActivityDetailResp.java
@@ -112,7 +112,7 @@ public class SeckillActivityDetailResp implements Serializable {
         private Integer itemQuantity;
 
         @ApiModelProperty("商品原价格")
-        private Long originalPrice;
+        private BigDecimal originalPrice;
     }
 
     @Data
